Simplified roll button creation and added saving throws
Previously I created each button individually and saved them to an array. This array was never used since the buttons have no input and only need to be connected to a function. I now just read from a list of the attribute names.
Saving throws have been implemented. Using the recalculate and the fix_labels functions the modifiers for the saving throws are kept up to date during proficiency changes, attribute score changes, or changing if the save is proficient.