Modified roll output and fixed some bugs
Roll output is now more clean.

Program would crash when canceling while saving or loading and this was fixed by returning near the begining of these functions.
Also noticed that "Performance" checkbox for proficiency was not responding on left side. This was because the size of the label widget for passive percepton was too large.